Fixed combinedImageStack bug and scattering
I found a bug in the create_total_average.m file. Because my naming scheme was not unique, when the normalization to the max of a replicate was done, also the images from the _myfilt_nostedfilt, were incorporated. They were however not incorporated in the CombinedImageStack, because there they get filtered out due to the classification selection.

Also scattering had some bugs, cytosolic and membrane proteins were allowed to be place also in both compartments. E.g. membrane protein could be placed in the cytosol.